py systems: Split and hoist general header utilities
