+++ bindings/pydrake/common/BUILD.bazel
@@ -103,6 +103,31 @@ drake_cc_library(
 )
 
 drake_cc_library(
+    name = "default_scalars_pybind",
+    hdrs = ["default_scalars_pybind.h"],
+    declare_installed_headers = 0,
+    visibility = ["//visibility:public"],
+    deps = [
+        ":type_pack",
+        "//bindings/pydrake:autodiff_types_pybind",
+        "//bindings/pydrake:pydrake_pybind",
+        "//bindings/pydrake:symbolic_types_pybind",
+    ],
+)
+
+# N.B. Any C++ libraries that include this must include `cpp_template_py` when
+# being used in Python.
+drake_cc_library(
+    name = "value_pybind",
+    hdrs = ["value_pybind.h"],
+    declare_installed_headers = 0,
+    visibility = ["//visibility:public"],
+    deps = [
+        "//bindings/pydrake/common:cpp_template_pybind",
+    ],
+)
+
+drake_cc_library(
     name = "eigen_pybind",
     hdrs = ["eigen_pybind.h"],
     declare_installed_headers = 0,
