+++ bindings/pydrake/systems/BUILD.bazel
@@ -40,18 +40,15 @@ drake_py_library(
     ],
 )
 
-# N.B. Any C++ libraries that include this must include `cpp_template_py` when
-# being used in Python.
 drake_cc_library(
     name = "systems_pybind",
     hdrs = ["systems_pybind.h"],
     declare_installed_headers = 0,
     visibility = ["//visibility:public"],
     deps = [
-        "//bindings/pydrake:autodiff_types_pybind",
-        "//bindings/pydrake:symbolic_types_pybind",
-        "//bindings/pydrake/common:cpp_param_pybind",
-        "//bindings/pydrake/common:cpp_template_pybind",
+        "//bindings/pydrake:pydrake_pybind",
+        "//bindings/pydrake/common:default_scalars_pybind",
+        "//bindings/pydrake/common:value_pybind",
     ],
 )
